requirements.txt
discord.py>=1.7.1
python-dotenv>=0.17.1
jishaku>=1.20.0.220
pytz>=2021.1